# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DiscoveryPage": {
        "additionalProperties": false,
        "description": "One page of discovery results.\n\nCarries `total` because `CLAUDE.md` section 6 asks for consistent list\nconventions and no client can render \"page N of M\" without it. Noted in\nTASKS.md as the shape the other list endpoints should be retrofitted to.",
        "example": {
          "items": [],
          "limit": 20,
          "offset": 0,
          "total": 0
        },
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StartupCard"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total matching startups, ignoring pagination.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "DiscoveryPage",
        "type": "object"
      },
...
      "Stage": {
        "description": "Funding stage.\n\nA bounded vocabulary, unlike sector: benchmarks are keyed by stage, so it\nhas to match across startups to be comparable at all.",
        "enum": [
          "idea",
          "pre_seed",
          "seed",
          "series_a",
          "series_b_plus",
          "growth"
        ],
        "title": "Stage",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "description": "Access token issued by `POST /v1/auth/login`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for the FundReady platform. All resources live under `/v1`. Timestamps are ISO 8601 UTC, currencies are ISO 4217 codes, and money is expressed in integer minor units.\n\nEvery error response uses the envelope `{\"error\": {\"code\", \"message\", \"details\"}}`. Branch on `code`, which is stable; never parse `message`.\n\nEvery response carries an `X-Request-ID` header. Quote it in a support report and we can find the exact log line.",
    "summary": "AI business audit and investor-matching platform for SACI Holdings.",
    "title": "FundReady API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/discover": {
      "get": {
        "description": "Startups that have published themselves, newest first.\n\nFilters are the four indexed columns. `sector` matches case-insensitively because sector is free text \u2014 a founder who typed `Fintech` and a filter of `fintech` mean the same thing.\n\n`total` is the count ignoring pagination, so a client can render \"page N of M\".\n\nA startup appears only once it has a **succeeded** audit. A published profile with no verdict is not listed: a card with nothing on it invites a direct approach, which is what the brokerage prevents.\n\n**Investors and SACI admins only; `403` for a founder.**\n\nEvery startup here opted in explicitly \u2014 running an audit does not make a founder discoverable, publishing does. Results carry the **summary tier** only: the two verdicts and the four discovery columns. There is no founder name, no contact detail, and none of the submitted figures. That is the brokerage: the verdict is enough to decide whether to ask for an introduction, and not enough to skip one.",
        "operationId": "discover_v1_discover_get",
        "parameters": [
          {
            "in": "query",
            "name": "sector",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Stage"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Browse discoverable startups",
        "tags": [
          "investor"
        ]
      }
    },
...
    "/v1/documents/{document_id}/complete": {
      "post": {
        "description": "Call this after the `PUT` succeeds. The server reads the object back from storage and judges it on what actually arrived -- not on what you declared.\n\nA file that is missing, empty, larger than `max_bytes`, or of an unaccepted type is **deleted from storage** and the document is marked `rejected`; `422` carries the reason in `details.reason` (`object_missing`, `empty_file`, `file_too_large`, `unsupported_content_type`).\n\nSafe to retry: calling it again on a `ready` document returns it unchanged.\n\nYou can only reach your own profile. Another founder's id returns `404`, not `403` -- deliberately, so the API does not confirm which ids exist. SACI admins can read any profile.",
        "operationId": "complete_upload_v1_documents__document_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Confirm a document upload",
        "tags": [
          "intake"
        ]
      }
    },
    "/v1/documents/{document_id}/download": {
      "get": {
        "description": "Returns a short-lived URL, not the file. `GET` it directly with no `Authorization` header.\n\n**Treat the URL as a credential.** Anyone holding it can read the file until it expires -- do not log it or persist it.\n\n`422` if the document is not `ready`, or if a scan has flagged it.\n\nYou can only reach your own profile. Another founder's id returns `404`, not `403` -- deliberately, so the API does not confirm which ids exist. SACI admins can read any profile.",
        "operationId": "download_document_v1_documents__document_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadTicket"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get a download URL for a document",
        "tags": [
          "intake"
        ]
      }
    },
...
    "/v1/registries": {
      "get": {
        "description": "Static reference data for the registration form: given an ISO 3166-1 alpha-2 country code, which body registers companies there, what the certificate is called, and how to label the registration number field.\n\n**Key on ISO codes, not display names.** `NG` maps to CAC; `\"Nigeria\"` is not a key. Countries not in this list are still accepted \u2014 the founder fills legal name and number as free text and uploads a `registration_certificate`.\n\nNothing here verifies a company exists (`DECISIONS.md` D7).",
        "operationId": "list_registries_v1_registries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List company registries by country",
        "tags": [
          "intake"
        ]
      }
    },
    "/v1/startups": {
      "post": {
        "description": "Creates the profile for the authenticated founder. **Every field is optional** -- start with what you know and fill the rest in later, or let document extraction do it. `missing_fields` on the response says what an audit will still need.\n\nThe owner is taken from your token, never from the request body.\n\n**If you omit `name`, it is filled in from your company email domain** -- `founder@acme.com` gives `Acme`. It is a starting point, not a verified company name: send `name` yourself to set it exactly, or correct it later with `PATCH /startups/{id}`. A name you do send is never overwritten.\n\nFounders (and SACI admins) only. `403` for an investor.\n\n`409` if you already have a profile: one per founder in v1.",
        "operationId": "create_profile_v1_startups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartupProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create your startup profile",
        "tags": [
          "intake"
        ]
      }
    },
    "/v1/startups/me": {
      "get": {
        "description": "Returns the authenticated founder's profile without needing its id.\n\n`404` if you have not created one yet.",
        "operationId": "read_own_profile_v1_startups_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Your own startup profile",
        "tags": [
          "intake"
        ]
      }
    },
...
    "/v1/startups/{profile_id}/publish": {
      "post": {
        "description": "Opt this startup in to investor discovery.\n\nIdempotent: publishing an already-published profile refreshes `published_at` and changes nothing else.\n\n**Publishing is consent, not eligibility.** Opting in says the founder is willing to be seen. Whether there is anything to *show* is decided separately: a startup appears in `GET /v1/discover` only once it also has a **succeeded audit**, because a card with no verdict tells an investor nothing and invites them to approach the founder directly \u2014 which is what the brokerage exists to prevent.\n\nSo publishing before the first audit finishes is allowed and is not an error. The client should say so: *\"You'll appear to investors once your audit completes.\"* Poll the audit, not this endpoint.\n\n**What an investor then sees is the summary tier only** \u2014 the two verdict levels and scores, plus sector, stage and country. Never the submitted figures, the findings, the action plan, or any contact detail. The full report is revealed by a SACI admin at the meeting and by no other path.\n\nYou can only reach your own profile. Another founder's id returns `404`, not `403` -- deliberately, so the API does not confirm which ids exist. SACI admins can read any profile.",
        "operationId": "publish_profile_v1_startups__profile_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Publish to investors",
        "tags": [
          "intake"
        ]
      }
    },
...
  }
}